move `CommentLoader` to data source
